% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_e_alpha}
\alias{WH_e_alpha}
\title{Critical value for Nair's equal-precision confidence bands (Borokov–Sycheva)}
\usage{
WH_e_alpha(lower, upper, alpha, tol = 1e-10, maxit = 10000L)
}
\arguments{
\item{lower}{Lower limit of supremum evaluation.}

\item{upper}{Upper limit of supremum evaluation.}

\item{alpha}{Significance level.}

\item{tol}{Convergence tolerance. Default = \code{1e-10}.}

\item{maxit}{Maximum number of iterations. Default = \code{10000}.}
}
\value{
The critical value \emph{e}.
}
\description{
Computes the critical value \emph{e} for Nair's equal-precision confidence bands¹
using Borokov–Sycheva approximation². Note: to reproduce Table 2 of Nair
(1982), use \code{alpha/2}.
}
\references{
\enumerate{
\item Nair, V.N., 1984. Conﬁdence bands for survival functions with censored
data: a comparative study. \emph{Technometrics}, 26, pp. 265–275.
\item Borokov, A.A. and Sycheva, N.M., 1968. On asymptotically optimal
non-parametric criteria. \emph{Theory of Probability & Its Applications}, 13(3),
pp. 359–393.
}
}
