% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_e_alpha_MC}
\alias{WH_e_alpha_MC}
\title{Critical value for Nair's equal-precision confidence bands (Monte Carlo)}
\usage{
WH_e_alpha_MC(
  lower,
  upper,
  alpha,
  n_step = 100000L,
  n_rep = 100000L,
  seed = 24601L
)
}
\arguments{
\item{lower}{Lower limit of supremum evaluation.}

\item{upper}{Upper limit of supremum evaluation.}

\item{alpha}{Significance level.}

\item{n_step}{Number of discretization steps for each variance-stabilized
Brownian bridge. Default = \code{1e5}.}

\item{n_rep}{Number of Monte Carlo replicates. Default = \code{1e5}.}

\item{seed}{Random seed. Default = \code{24601}.}
}
\value{
The critical value \eqn{e_\alpha}.
}
\description{
Computes the critical value \emph{e} for Nair's equal-precision confidence bands¹
using Monte Carlo simulation. Note: to reproduce Table 2 of Nair (1984), use
\code{alpha/2}.
}
\details{
Parallelized using OpenMP if available.
}
\references{
\enumerate{
\item Nair, V.N., 1984. Conﬁdence bands for survival functions with censored
data: a comparative study. \emph{Technometrics}, 26, pp. 265–275.
}
}
