% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_HollanderMcKeague}
\alias{WH_HollanderMcKeague}
\title{Hollander–McKeague confidence bands}
\usage{
WH_HollanderMcKeague(
  time,
  risk,
  event,
  alpha = 0.05,
  verbose = TRUE,
  tol_G = 1e-10,
  maxit_G = 10000L,
  tol_K = 1e-10,
  maxit_K = 10000L
)
}
\arguments{
\item{time}{Numeric vector of unique event times.}

\item{risk}{Integer vector of numbers at risk at each \code{time}.}

\item{event}{Integer vector of numbers of events at each \code{time}.}

\item{alpha}{Optional significance level. Default = \code{0.05}.}

\item{verbose}{Optional logical; if \code{TRUE}, prints additional diagnostic
information. Default = \code{TRUE}.}

\item{tol_G}{Optional convergence tolerance for the infinite series
computing the Hall–Wellner² distribution function \emph{G}(\emph{a}, \emph{λ}). Default =
\code{1e-10}.}

\item{maxit_G}{Optional maximum number of iterations for the infinite series
computing the Hall–Wellner distribution function \emph{G}(\emph{a}, \emph{λ}). Default =
\code{10000}.}

\item{tol_K}{Optional convergence tolerance for the bisection root-finding
of the Hall–Wellner quantile function \emph{K}(\emph{x}, \emph{a}). Default = \code{1e-10}.}

\item{maxit_K}{Optional maximum number of iterations for the bisection
root-finding of the Hall–Wellner quantile function \emph{K}(\emph{x}, \emph{a}). Default =
\code{10000}.}
}
\value{
A numeric matrix with 2 columns: lower limit and upper limit.
}
\description{
Computes the Hollander–McKeague likelihood-ratio simultaneous confidence
band¹ for the Kaplan–Meier survival estimator at each unique event time.
}
\details{
The Hollander–McKeague procedure begins with a specified \emph{simultaneous} \ifelse{latex}{\out{$\mathit{\alpha}$}}{\ifelse{html}{\out{<i>&alpha;</i>}}{*alpha*}} and
computes a smaller \emph{pointwise} \ifelse{latex}{\out{$\mathit{\alpha}$}}{\ifelse{html}{\out{<i>&alpha;</i>}}{*alpha*}} such that the resulting \emph{pointwise} confidence
intervals achieve the desired \emph{simultaneous} coverage. Extremely small
\emph{pointwise} \ifelse{latex}{\out{$\mathit{\alpha}$}}{\ifelse{html}{\out{<i>&alpha;</i>}}{*alpha*}} can cause 1 − \ifelse{latex}{\out{$\mathit{\alpha}$}}{\ifelse{html}{\out{<i>&alpha;</i>}}{*alpha*}} to round to 1 in double precision. To prevent
this, \emph{pointwise} \ifelse{latex}{\out{$\mathit{\alpha}$}}{\ifelse{html}{\out{<i>&alpha;</i>}}{*alpha*}} is rounded up to at least half the machine epsilon.
}
\references{
\enumerate{
\item Hollander, M. and McKeague, I.W., 1997. Likelihood ratio-based confidence
bands for survival functions. \emph{Journal of the American Statistical
Association}, 92(437), pp. 215–226.
\item Hall, W.J. and Wellner, J.A., 1980. Confidence bands for a survival curve
from censored data. \emph{Biometrika}, 67(1), pp. 133–143.
}
}
