% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_HollanderMcKeague_alpha}
\alias{WH_HollanderMcKeague_alpha}
\title{Hollander–McKeague pointwise significance level}
\usage{
WH_HollanderMcKeague_alpha(
  risk,
  event,
  surv,
  SE,
  alpha = 0.05,
  tol_G = 1e-10,
  maxit_G = 10000L,
  tol_K = 1e-10,
  maxit_K = 10000L
)
}
\arguments{
\item{risk}{Integer vector of numbers at risk at each unique event time.}

\item{event}{Integer vector of numbers of events at each unique event time.}

\item{surv}{Numeric vector of Kaplan–Meier survival estimates at each unique
event time.}

\item{SE}{Numeric vector of standard errors for each \code{surv}.}

\item{alpha}{Optional significance level. Default = \code{0.05}.}

\item{tol_G}{Optional convergence tolerance for the infinite series
computing the Hall–Wellner² distribution function \emph{G}(\emph{a}, \emph{λ}). Default =
\code{1e-10}.}

\item{maxit_G}{Optional maximum number of iterations for the infinite series
computing the Hall–Wellner distribution function \emph{G}(\emph{a}, \emph{λ}). Default =
\code{10000}.}

\item{tol_K}{Optional convergence tolerance for the bisection root-finding
of the Hall–Wellner quantile function \emph{K}(\emph{x}, \emph{a}). Default = \code{1e-10}.}

\item{maxit_K}{Optional maximum number of iterations for the bisection
root-finding of the Hall–Wellner quantile function \emph{K}(\emph{x}, \emph{a}). Default =
\code{10000}.}
}
\value{
The pointwise significance level \ifelse{latex}{\out{$\mathit{\alpha}_{\mathit{p}}$}}{\ifelse{html}{\out{<i>&alpha;</i><sub><i>p</i></sub>}}{*alpha_p*}}.
}
\description{
Computes the \emph{pointwise} \ifelse{latex}{\out{$\mathit{\alpha}_{\mathit{p}}$}}{\ifelse{html}{\out{<i>&alpha;</i><sub><i>p</i></sub>}}{*alpha_p*}} for Hollander–McKeague confidence bands.
}
\references{
\enumerate{
\item Hollander, M. and McKeague, I.W., 1997. Likelihood ratio-based confidence
bands for survival functions. \emph{Journal of the American Statistical
Association}, 92(437), pp. 215–226.
\item Hall, W.J. and Wellner, J.A., 1980. Confidence bands for a survival curve
from censored data. \emph{Biometrika}, 67(1), pp. 133–143.
}
}
