% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_KMconf_med}
\alias{WH_KMconf_med}
\title{Kaplan–Meier median survival confidence limits}
\usage{
WH_KMconf_med(input)
}
\arguments{
\item{input}{A numeric matrix with 3 columns: time, lower limit, and upper
limit.}
}
\value{
A numeric matrix with 2 columns: lower limit and upper limit. \code{NaN}
indicates that the median is not reached.
}
\description{
Confidence limits for median survival based on the times at which the
confidence limits of the survival function cross 0.5¹.
}
\references{
\enumerate{
\item Brookmeyer, R. and Crowley, J., 1982. A confidence interval for the
median survival time. \emph{Biometrics}, pp. 29–41.
}
}
