% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_Rothman}
\alias{WH_Rothman}
\title{Rothman confidence intervals}
\usage{
WH_Rothman(time, surv, risk, event, alpha = 0.05)
}
\arguments{
\item{time}{Numeric vector of unique event times.}

\item{surv}{Numeric vector of Kaplan–Meier survival estimates at each \code{time}.}

\item{risk}{Integer vector of numbers at risk at each \code{time}.}

\item{event}{Integer vector of numbers of events at each \code{time}.}

\item{alpha}{Optional significance level. Default = \code{0.05}.}
}
\value{
A numeric matrix with 3 columns: time, lower limit, and upper limit.
}
\description{
Computes Rothman binomial pointwise confidence intervals¹ for the
Kaplan–Meier survival estimator at each unique event time.
}
\references{
\enumerate{
\item Rothman, K.J., 1978. Estimation of confidence limits for the cumulative
probability of survival in life table analysis. \emph{Journal of Chronic Diseases},
31(8), pp. 557–560.
}
}
