% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_ThomasGrunkemeier}
\alias{WH_ThomasGrunkemeier}
\title{Thomas–Grunkemeier confidence intervals}
\usage{
WH_ThomasGrunkemeier(time, risk, event, alpha = 0.05)
}
\arguments{
\item{time}{Numeric vector of unique event times.}

\item{risk}{Integer vector of numbers at risk at each \code{time}.}

\item{event}{Integer vector of numbers of events at each \code{time}.}

\item{alpha}{Significance level. Default = \code{0.05}.}
}
\value{
A numeric matrix with 2 columns: lower limit and upper limit.
}
\description{
Computes Thomas–Grunkemeier likelihood-ratio pointwise confidence intervals
for the Kaplan–Meier survival estimator at each unique event time.
}
