% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_HollanderMcKeague}
\alias{WH_HollanderMcKeague}
\title{Hollander–McKeague confidence bands}
\usage{
WH_HollanderMcKeague(
  time,
  risk,
  event,
  alpha = 0.05,
  warn = TRUE,
  tol_G = 1e-10,
  maxit_G = 10000L,
  tol_K = 1e-10,
  maxit_K = 10000L
)
}
\arguments{
\item{time}{Numeric vector of unique event times.}

\item{risk}{Integer vector of numbers at risk at each \code{time}.}

\item{event}{Integer vector of numbers of events at each \code{time}.}

\item{alpha}{Significance level. Default = \code{0.05}.}

\item{warn}{Logical; if \code{TRUE}, prints a message if the \emph{pointwise α} is
rounded up.}

\item{tol_G}{Optional convergence tolerance for the infinite series
computing the Hall–Wellner distribution function \emph{G}(\emph{a}, \emph{λ}). Default =
\code{1e-10}.}

\item{maxit_G}{Optional maximum number of iterations for the infinite series
computing the Hall–Wellner distribution function \emph{G}(\emph{a}, \emph{λ}). Default =
\code{10000}.}

\item{tol_K}{Optional convergence tolerance for the bisection root-finding
of the Hall–Wellner quantile function \emph{K}(\emph{x}, \emph{a}). Default = \code{1e-10}.}

\item{maxit_K}{Optional maximum number of iterations for the bisection
root-finding of the Hall–Wellner quantile function \emph{K}(\emph{x}, \emph{a}). Default =
\code{10000}.}
}
\value{
A numeric matrix with 2 columns: lower limit and upper limit.
}
\description{
Computes the Hollander–McKeague likelihood-ratio simultaneous confidence
band for the Kaplan–Meier survival estimator at each unique event time.
}
\details{
The Hollander–McKeague procedure begins with a specified \emph{simultaneous α} and
computes a smaller \emph{pointwise α} such that the resulting \emph{pointwise} confidence
intervals achieve the desired \emph{simultaneous} coverage. Extremely small
\emph{pointwise α} can cause 1 − \emph{α} to round to 1 in double precision. To prevent
this, \emph{pointwise α} is rounded up to at least half the machine epsilon.
}
