% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_Rothman}
\alias{WH_Rothman}
\title{Rothman confidence intervals}
\usage{
WH_Rothman(surv, risk, event, alpha = 0.05)
}
\arguments{
\item{surv}{Numeric vector of Kaplan–Meier survival probabilities.}

\item{risk}{Integer vector of numbers at risk at each change in \code{surv}.}

\item{event}{Integer vector of event counts at each change in \code{surv}.}

\item{alpha}{Significance level. Default = \code{0.05}.}
}
\value{
A numeric matrix with 2 columns: lower limit and upper limit.
}
\description{
Computes Rothman binomial pointwise confidence intervals for the
Kaplan–Meier survival estimator at each unique event time.
}
